requirements.txt
boto3
requests
python-docx
PyPDF2
openpyxl
tiktoken
cohere
